add Function to scenario batch remove from scenario execution
